Changed Integer/Long to double. Changed some API response key names.

+++ src/main/java/com/metric/oyometrics_team2/DTO/FirstCommentResponse.java
@@ -12,5 +12,5 @@ import org.springframework.stereotype.Component;
 @ToString
 @Component
 public class FirstCommentResponse {
-    private Long firstCommentMeanTime;
+    private double firstCommentMeanTime;
 }
